Product admin securing the view

+++ products/views.py
@@ -4,6 +4,9 @@ from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
+
+
 
 # Create your views here.
 def all_products(request):
@@ -74,9 +77,14 @@ def product_detail(request, product_id):
     
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admin user can do that.')
+        return redirect(reverse('home'))
+
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -96,9 +104,13 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+     if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admin user can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -121,9 +133,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
-
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+     if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admin user can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Successfully deleted the {product.name} from the store')
